pyproject.toml
[project]
name = "qc-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite~=0.22.1",
    "alembic~=1.18.4",
    "arize-phoenix~=8.32.1",
    "asyncio-mqtt~=0.16.2",
    "bcrypt==4.3.0",
    "chromadb~=1.5.7",
    "faker~=40.13.0",
    "fastapi~=0.135.3",
    "httpx~=0.28.1",
    "langchain~=1.2.15",
    "langchain-community~=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai~=1.1.12",
    "langgraph~=1.1.6",
    "numpy~=2.4.4",
    "openai~=2.31.0",
    "openinference-instrumentation-langchain~=0.1.62",
    "pandas~=3.0.2",
    "passlib[bcrypt]~=1.7.4",
    "plotly~=6.6.0",
    "pydantic-settings~=2.13.1",
    "pydantic[email]~=2.12.5",
    "python-dotenv~=1.2.2",
    "python-jose[cryptography]~=3.5.0",
    "python-multipart~=0.0.24",
    "ragas~=0.4.3",
    "sqlalchemy~=2.0.49",
    "sqlean-py==3.49.1",
    "streamlit~=1.56.0",
    "streamlit-autorefresh~=1.0.1",
    "uvicorn[standard]~=0.44.0",
    "websocket-client~=1.9.0",
    "websockets~=16.0",
]
